backend/requirements.txt
Click==7.0
Flask==1.0.3
flask-heroku==0.1.9
Flask-SQLAlchemy==2.4.0
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2==2.8.2
PyJWT==1.7.1
SQLAlchemy==1.3.3
Werkzeug==0.15.4
